Port lock.go to Windows
